<Refactor> 이미지 압축 코드 분리
